% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_NVAR}
\alias{sim_NVAR}
\title{Time series simulation with an NVAR model}
\usage{
sim_NVAR(model, init)
}
\arguments{
\item{model}{An \code{NVAR} model, fitted by \code{\link[=NVAR]{NVAR()}}.}

\item{init}{A \code{tibble}, data.frame, or matrix that specify the initial values for a simulation. Should contain the variables used to fit the model and be at least \eqn{s * (k - 1)} long.}
}
\description{
Time series simulation with an NVAR model
}
