% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVAR.R
\name{NVAR}
\alias{NVAR}
\title{Fit a nonlinear vector autoregression model}
\usage{
NVAR(data, s, k, p, ...)
}
\arguments{
\item{data}{A matrix or data frame that represents a time series of vectors, with each row as a time step.}

\item{s}{The number of time steps used for constructing features.}

\item{k}{The number of time steps skipped between each two used time steps.}

\item{p}{The order of polynomial feature vector.}

\item{...}{Other parameters passed to \code{\link[glmnet:glmnet]{glmnet::glmnet()}}}
}
\description{
Described by Gauthier et al. (2021), also known as the "next generation reservoir computing" (NG-RC).
}
\details{
The feature vector is as follows (from the reference):

\deqn{\mathbb{O}_{\text {total }}=\mathbb{O}_{\text {lin }} \oplus \mathbb{O}_{\text {nonlinear }}^{(p)}}

\deqn{\mathbb{O}_{\operatorname{lin}, i}=\mathbf{X}_i \oplus \mathbf{X}_{i-s} \oplus \mathbf{X}_{i-2 s} \oplus \ldots \oplus \mathbf{X}_{i-(k-1) s}}

\deqn{\mathbb{O}_{\text {nonlinear }}^{(p)}=\mathbb{O}_{\text {lin }}\lceil\otimes\rceil \mathbb{O}_{\text {lin }}\lceil\otimes\rceil \ldots\lceil\otimes\rceil \mathbb{O}_{\text {lin }}}

The feature vector \eqn{\mathbb{O}_{\text {total }}} is then used as input for a ridge regression.
}
\references{
Gauthier, D. J., Bollt, E., Griffith, A., & Barbosa, W. A. S. (2021). Next generation reservoir computing. Nature Communications, 12(1), 5564. https://doi.org/10.1038/s41467-021-25801-2
}
